feat: move bench() method and bench test proc macro to toools/benches
