+++ kclvm/tools/src/LSP/Cargo.toml
@@ -41,4 +41,4 @@ salsa = { version = "0.16.1", default-features = false }
 serde_json = { version = "1.0", default-features = false }
 parking_lot = { version = "0.12.0", default-features = false }
 rustc-hash = { version = "1.1.0", default-features = false }
-proc_macro_crate = {path = "./proc_macro_crate"}
+proc_macro_crate = {path = "../../benches/proc_macro_crate"}
