+++ kclvm/tools/src/LSP/src/lib.rs
@@ -16,4 +16,3 @@ mod goto_def;
 mod hover;
 mod quick_fix;
 mod request;
-mod test_utils;
